submission/supplementary/code/core/field.py
"""Unified field theory implementation."""

from typing import Dict, Optional, Union, List
import numpy as np
from sympy import Symbol, exp, integrate
from .physics_constants import (
    ALPHA_VAL, X, E, Z_MASS,
    g1_REF, g2_REF, g3_REF,
    ALPHA_REF, GAMMA_1, GAMMA_2, GAMMA_3
)
from .basis import FractalBasis
from .types import Energy, FieldConfig, WaveFunction
from .modes import ComputationMode
from .errors import PhysicsError, ValidationError

class UnifiedField:
    """Implementation of unified field theory."""
    
    def __init__(self, alpha: float = ALPHA_VAL, mode: ComputationMode = ComputationMode.SYMBOLIC):
        """Initialize unified field."""
        self.alpha = alpha
        self.mode = mode
        self.basis = FractalBasis(alpha=alpha, mode=mode)
        
    def compute_field(self, config: FieldConfig) -> WaveFunction:
        """
        Compute field configuration.
        
        Args:
            config: Field configuration parameters
            
        Returns:
            WaveFunction: Computed field configuration
        """
        raise NotImplementedError("Field computation not implemented")
        
    def evolve(self, energy: Energy) -> WaveFunction:
        """
        Evolve field to given energy scale.
        
        Args:
            energy: Target energy scale
            
        Returns:
            WaveFunction: Evolved field configuration
        """
        raise NotImplementedError("Field evolution not implemented")